refactor: 컨밴션에 맞추어 파일 이름에 .component 제거
